feat : 패키지 분리, 추가 테스트, 중복 제거 기능 개발
