+++ chess/src/main/java/chess/Board.java
@@ -1,0 +1,29 @@
+package chess;
+
+import java.util.ArrayList;
+
+public class Board {
+
+    ArrayList<Pawn> boardList = new ArrayList<>();
+
+    public void add(Pawn pawn) {
+        boardList.add(pawn);
+    }
+
+    public int size() {
+        return boardList.size();
+    }
+
+    // 인덱스를 사용하여 Pawn 객체 찾기
+    public Pawn findPawn(int idx) {
+        if (isValidIndex(idx)) {
+            return boardList.get(idx);
+        }
+        return null;
+    }
+
+    //유효성 검사
+    private boolean isValidIndex(int idx) {
+        return idx >= 0 && idx < boardList.size();
+    }
+}
