+++ chess/src/main/java/chess/Main.java
@@ -1,4 +1,4 @@
-package org.example;
+package chess;
 
 //TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
 // click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
